+++ tomcat/src/main/java/org/apache/coyote/http11/controller/impl/IndexController.java
@@ -1,7 +1,5 @@
-package nextstep.jwp.controller;
+package org.apache.coyote.http11.controller.impl;
 
-import nextstep.jwp.db.InMemoryUserRepository;
-import nextstep.jwp.model.User;
 import org.apache.coyote.http11.ContentType;
 import org.apache.coyote.http11.HttpHeaders;
 import org.apache.coyote.http11.HttpRequest;
@@ -13,13 +11,7 @@ import org.apache.coyote.http11.controller.AbstractController;
 import java.nio.file.Files;
 import java.nio.file.Path;
 
-public final class RegisterController extends AbstractController {
-
-    @Override
-    protected void doPost(HttpRequest request, HttpResponse response) throws Exception {
-        InMemoryUserRepository.save(new User(request.getRequestBody().get("account"), request.getRequestBody().get("password"), request.getRequestBody().get("email")));
-        response.sendRedirect("/index.html");
-    }
+public final class IndexController extends AbstractController {
 
     @Override
     protected void doGet(HttpRequest request, HttpResponse response) throws Exception {
